+++ test/Armada.Test.Unit/Program.cs
@@ -349,6 +349,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new BuildDriftServiceTests());
             runner.AddSuite(new SelfDeployServiceTests());
             runner.AddSuite(new DatabaseBackupServiceTests());
+            runner.AddSuite(new SelfDeployNativeClientTests());
             runner.AddSuite(new SelfDeployDatabaseBackupProviderTests());
             runner.AddSuite(new SelfDeployBuildRunnerTests());
             runner.AddSuite(new SelfDeployNativeIntegrationTests());